rewrite spawn to use varargs
it's seems neater than packing an array on the spot, we just do that
inside spawn before calling execvp.  If we get too many arguments, just
abort.

This also changes how the forking is handled: I set RVSEL only for the
child and call disable_handlers() and endwin() preemptively to avoid
possible races with our children.